requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2

openai==0.27.6
python-dotenv==1.0.0
urllib3==1.25.7
DateTime==5.1